LayoutTests/js/dom/document-all-watchpoint-covers-eliminated-compare-eq-expected.txt

Test to make sure that document.all works correctly with elminated CompareEq in DFG.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS documentAllCompare.isNull is true
PASS documentAllCompare.isUndefined is true
PASS documentAllCompare.length is 13
PASS successfullyParsed is true

TEST COMPLETE
